requirements.txt
apache-airflow
python-dotenv
pandas
beautifulsoup4
psycopg2-binary